[#147] refact: 헤더 서버사이드로 분리
